perf: Use double-buffering for spans
Previously we would drain the collector buffer into an exporter buffer,
which would be O(n), block the mutex lock for that period of time, and
re-allocate a fresh Vec every time (though presumably in one step, as
Drain implements ExactSizeIterator).

With this double-buffering approach there are two benefits:

1. We don't move around the heap allocation anymore, mem::swap is O(1),
   which also means we hold the lock for much shorter.
2. We keep both buffers' capacity allocated, so that after the initial
   two rounds, they should continue running at the same volume without
   any further allocations.

A potential downside here is that a single iteration with outsized
volume will permanently grow a buffer accordingly, holding onto memory
it doesn't need. In the future we could potentially shrink buffers if
they're way larger than needed.