feat: Add support for 128-bit trace IDs
The main difficulty here was that the agent trace API only accepts 64
bits, and stores the upper 64 bits in a span tag instead. For span
links, it similarly splits it across two 64-bit fields.

This means we need custom serializers for both types, and I had to work
around serde a bit to add an entry to the span tags during serialization
without constructing a whole new HashMap.